+++ docs/md/Version-history.md
@@ -17,6 +17,7 @@ Available in [pre-release](https://www.sumatrapdfreader.org/prerelease) builds.
 - improve `Toggle Use Tabs` (`CmdToggleUseTabs`). You can now transition between using tabs / not using tabs witout restarting the app
 - allow showing menu bar when using tabs (previously menu bar was only shown when not using tabs)
 - add `CmdScreenshot` (`Take Screenshot` in `Ctrl + k` [command palette](Command-Palette.md)) to capture screenshots of the desktop and all visible windows, saved as PNG files in `Screenshots` sub-directory of SumatraPDF data directory
+- you can drag&drop images from a browser onto SumatraPDF window. We'll download it to Downloads folder and open
 
 ## 3.6 (2026-03-17)